// src/app/models/tema.ts
export interface ITema {
  PageIndex?: number;
  PageSize?: number;
  orderBy0?: number;
  orderByDirection0?: number;
  Tema_Titulo?: string;
  Tema_Descripcion?: string;
  Tema_Duracion_Minutos?: number;
  Registro_Usuario?: string;
  Tema_Numero?: number;

  Linea?: number;
  Ultima_Linea?: number;
  Cantidad_Registros?: number;
}